+++ src/Instance/ItemsList/ItemsList.js
@@ -3,6 +3,7 @@ import React, {
   useCallback,
   useState,
   useEffect,
+  useContext,
 } from 'react';
 import PropTypes from 'prop-types';
 import {
@@ -27,8 +28,12 @@ import {
 } from './utils';
 import useBoundWithHoldings from '../../Holding/ViewHolding/HoldingBoundWith/useBoundWithHoldings';
 
+import { DataContext } from '../../contexts';
+
 const getTableAria = (intl) => intl.formatMessage({ id: 'ui-inventory.items' });
 const getFormatter = (
+  intl,
+  locationsById,
   holding,
   holdingsMapById,
   selectItemsForDrag,
@@ -81,7 +86,15 @@ const getFormatter = (
   'copyNumber': ({ copyNumber }) => copyNumber || noValue,
   'materialType': x => x.materialType?.name || noValue,
   'loanType': x => x.temporaryLoanType?.name || x.permanentLoanType?.name || noValue,
-  'effectiveLocation': x => x.effectiveLocation?.name || noValue,
+  'effectiveLocation': x => {
+    const effectiveLocation = locationsById[x.effectiveLocation?.id];
+    return effectiveLocation?.isActive ?
+      effectiveLocation?.name || noValue :
+      intl.formatMessage(
+        { id: 'ui-inventory.inactive.gridCell' },
+        { location: effectiveLocation?.name }
+      );
+  },
   'enumeration': x => x.enumeration || noValue,
   'chronology': x => x.chronology || noValue,
   'volume': x => x.volume || noValue,
@@ -145,6 +158,7 @@ const ItemsList = ({
   });
   const [records, setRecords] = useState([]);
   const [paginatedItems, setPaginatedItems] = useState([]);
+  const { locationsById } = useContext(DataContext);
 
   const ariaLabel = useMemo(() => getTableAria(intl), []);
   const columnMapping = useMemo(
@@ -152,7 +166,7 @@ const ItemsList = ({
     [holding.id, records, isItemsDragSelected, selectItemsForDrag],
   );
   const formatter = useMemo(
-    () => getFormatter(holding, holdingsMapById, selectItemsForDrag, isItemsDragSelected),
+    () => getFormatter(intl, locationsById, holding, holdingsMapById, selectItemsForDrag, isItemsDragSelected),
     [holding, holdingsMapById, selectItemsForDrag, isItemsDragSelected],
   );
   const rowProps = useMemo(() => ({