+++ test/fixtures/holdingsRecords.js
@@ -1,0 +1,5 @@
+export const holdingsRecords = [
+  {
+    id: '44a21d17-a666-4f34-b24d-644f8ed1537b',
+  }
+];