+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Browse refactor: Show instance result in third pane when Number of titles = `1`. Refs UIIN-2186.
 * Optimistic locking error appears when user adds more than 1 tag to "Holdings" record. Fixes UIIN-2242.
 * Fix holdings view bound-with table also displaying directly linked items. Fixes UIIN-2260.
+* Display "Inactive" by inactive locations on instance view. Fixes UIIN-1970.
 * Create a new instance search option. Fixes UIIN-2264.
 
 ## [9.2.0](https://github.com/folio-org/ui-inventory/tree/v9.2.0) (2022-10-27)